chore(inao-cem): Refactor custom components and fetch tax data (#17195)
* chore: houskeeping, split up files and make code more radable

* chore: refactor out custom components

* fix: remove imports from codeowners

* chore: remove console.log

* chore: add loading state

* chore: simplify logic

* fix: coderabbit comments

* fix: match indentation to lint

* fix: codeowners

* chore: add space to match indentation

* chore: remove the inao folder from codeowners

* chore: .

* fix: typo in path

* fix: indentation

---------
