+++ libs/application/templates/inao/financial-statement-cemetery/src/dataProviders/index.ts
@@ -1,10 +1,7 @@
 import { defineTemplateApi } from '@island.is/application/types'
 import { UserProfileApi } from '@island.is/application/types'
-
-export {
-  NationalRegistryUserApi,
-  IdentityApi as IdentityApiProvider,
-} from '@island.is/application/types'
+import { IdentityApi } from '@island.is/application/types'
+export { NationalRegistryUserApi } from '@island.is/application/types'
 
 export const CurrentUserTypeProvider = defineTemplateApi({
   action: 'getUserType',
@@ -14,5 +11,11 @@ export const CurrentUserTypeProvider = defineTemplateApi({
 export const UserInfoApi = UserProfileApi.configure({
   params: {
     catchMock: true,
+  },
+})
+
+export const IdentityApiProvider = IdentityApi.configure({
+  params: {
+    includeActorInfo: true,
   },
 })
