+++ libs/application/templates/inao/financial-statement-cemetery/src/fields/CemeteryOverview/CapitalNumbersOverview.tsx
@@ -1,18 +1,19 @@
 import { Box, GridColumn, GridRow, Text } from '@island.is/island-ui/core'
-
-import { m } from '../../lib/messages'
 import { ValueLine } from './ValueLine'
 import { sectionColumn, starterColumnStyle } from './overviewStyles.css'
 import { useLocale } from '@island.is/localization'
-import { FinancialStatementCemetery } from '../../lib/dataSchema'
-import { formatCurrency } from '../../utils/helpers'
+import { FormValue } from '@island.is/application/types'
+import { getCapitalNumbersOverviewNumbers } from '../../utils/overviewUtils'
+import { m } from '../../lib/messages'
 
-export const CapitalNumberOverview = ({
-  answers,
-}: {
-  answers: FinancialStatementCemetery
-}) => {
+type Props = {
+  answers: FormValue
+}
+
+export const CapitalNumberOverview = ({ answers }: Props) => {
   const { formatMessage } = useLocale()
+  const { capitalIncome, capitalCost, totalCapital } =
+    getCapitalNumbersOverviewNumbers(answers)
   return (
     <>
       <Box className={starterColumnStyle}>
@@ -22,27 +23,17 @@ export const CapitalNumberOverview = ({
       </Box>
       <GridRow>
         <GridColumn span={['12/12', '6/12']} className={sectionColumn}>
-          <ValueLine
-            label={m.capitalIncome}
-            value={formatCurrency(answers.capitalNumbers.capitalIncome)}
-          />
+          <ValueLine label={m.capitalIncome} value={capitalIncome} />
         </GridColumn>
-        {answers.capitalNumbers?.capitalCost ? (
+        {capitalCost ? (
           <GridColumn span={['12/12', '6/12']} className={sectionColumn}>
-            <ValueLine
-              label={m.capitalCost}
-              value={formatCurrency(answers.capitalNumbers.capitalCost)}
-            />
+            <ValueLine label={m.capitalCost} value={capitalCost} />
           </GridColumn>
         ) : null}
       </GridRow>
       <GridRow>
         <GridColumn className={sectionColumn}>
-          <ValueLine
-            isTotal
-            label={m.totalCapital}
-            value={formatCurrency(answers.capitalNumbers?.total)}
-          />
+          <ValueLine isTotal label={m.totalCapital} value={totalCapital} />
         </GridColumn>
       </GridRow>
     </>
