+++ libs/application/templates/inao/financial-statement-cemetery/src/fields/index.ts
@@ -1,4 +1,3 @@
 export { CemeteryCaretaker } from './CemeteryCareteker'
-export { PowerOfAttorneyFields } from './PowerOfAttorney'
 export { CemeteryOverview } from './CemeteryOverview'
-export { CemeteryIncomeLimit } from './CemeteryIncomeLimit'
+export { FetchDataBasedOnSelectedYear } from './FetchDataBasedOnSelectedYear'
