--- libs/application/templates/inao/financial-statement-cemetery/src/forms/applicationForm/cemeteryCaretakerSection/caretakerMultiField.ts
@@ -1,17 +1,0 @@
-import { buildCustomField, buildMultiField } from '@island.is/application/core'
-import { m } from '../../../lib/messages'
-import { CEMETERYCARETAKER } from '../../../utils/constants'
-
-export const caretakerMultiField = buildMultiField({
-  id: 'caretakers',
-  title: m.cemeteryBoardmembers,
-  description: m.cemeteryRegisterCaretakers,
-  children: [
-    buildCustomField({
-      id: 'cemeteryCaretaker',
-      title: m.cemeteryBoardmembers,
-      component: 'CemeteryCaretaker',
-      childInputIds: Object.values(CEMETERYCARETAKER),
-    }),
-  ],
-})
