+++ libs/application/templates/inao/financial-statement-cemetery/src/utils/constants.ts
@@ -26,6 +26,7 @@ export enum TaxInfoTypes {
   CARE_INCOME = 300,
   BURIAL_REVENUE = 301,
   GRANT_FROM_THE_CEMETERY_FUND = 302,
+  DONATIONS_TO_CEMETERYFUND = 334,
 }
 
 export const CEMETERYOPERATIONIDS = {
